LUC-20 bug Cannot find local variable 'emprestimo' not resolve
